+++ src/main/java/com/example/Ecommerce/security/SecurityConfig.java
@@ -1,5 +1,8 @@
 package com.example.Ecommerce.security;
 
+import com.example.Ecommerce.security.jwt.JwtAuthenticationFilter;
+import com.example.Ecommerce.security.jwt.JwtTokenUtil;
+import com.example.Ecommerce.user.repository.LogoutAccessTokenRedisRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
