+++ src/main/java/com/example/Ecommerce/security/jwt/JwtAuthenticationFilter.java
@@ -1,5 +1,7 @@
-package com.example.Ecommerce.security;
+package com.example.Ecommerce.security.jwt;
 
+import com.example.Ecommerce.user.repository.LogoutAccessTokenRedisRepository;
+import com.example.Ecommerce.security.UserDetailServiceImpl;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
