+++ src/main/java/com/example/Ecommerce/security/jwt/JwtExpirationEnums.java
@@ -1,4 +1,4 @@
-package com.example.Ecommerce.security;
+package com.example.Ecommerce.security.jwt;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
