+++ src/main/java/com/example/Ecommerce/user/controller/UserController.java
@@ -1,6 +1,6 @@
 package com.example.Ecommerce.user.controller;
 
-import com.example.Ecommerce.security.JwtTokenUtil;
+import com.example.Ecommerce.security.jwt.JwtTokenUtil;
 import com.example.Ecommerce.security.UserDetailsImpl;
 import com.example.Ecommerce.user.dto.UserLoginDto;
 import com.example.Ecommerce.user.dto.UserRegisterDto;
