+++ src/main/java/com/example/Ecommerce/user/repository/LogoutAccessTokenRedisRepository.java
@@ -1,5 +1,6 @@
-package com.example.Ecommerce.security;
+package com.example.Ecommerce.user.repository;
 
+import com.example.Ecommerce.user.domain.LogoutAccessToken;
 import org.springframework.data.repository.CrudRepository;
 
 public interface LogoutAccessTokenRedisRepository extends CrudRepository<LogoutAccessToken, String> {
